requirements.txt

googleapiclient.discovery 
random
string
numpy
seaborn 
pandas 
dash
matplotlib.pyplot
plotly.graph_objs
plotly.express
openpyxl
random 
string
numpy 
seaborn 
pandas 
matplotlib.pyplot